Extract out stats collection class and test it
